refactor mtp
