+++ python/sglang/srt/speculative/spec_utils.py
@@ -54,8 +54,7 @@ def spec_need_hidden_states(server_args: Optional[ServerArgs] = None) -> bool:
     if server_args is None:
         server_args = get_global_server_args()
 
-    # TODO(lsyin): also skip when 1) step = 1 or 2) standalone draft model
-    return not server_args.enable_mtp
+    return False if server_args.speculative_algorithm == "MTP" else True
 
 
 @triton.jit
